Create API helper
This is a simple start of a helper where functions for making
common backend calls can be put. For example it could in the future
be extended with functionality for getting details about one single
car to display on the 'Learn' and 'Shop' pages, without having to
retrieve all car data from the backend every time.